# docs/openapi/api.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PenthouseLife Main API",
    "version": "1.0.0",
    "description": "REST API documentation for the PenthouseLife Main API"
  },
  "servers": [
    {
      "url": "/api",
      "description": "dev (this server)"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "description": "Paginated, newest first. Filter by role or search by name/email.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "description": "Matches name or email (case-insensitive)",
            "schema": {
              "type": "string"
            },
            "example": "priya"
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "BUYER",
                "AGENT",
                "ADMIN"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsers"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a user by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "clx1a2b3c0000abcd1234efgh"
          }
        ],
        "responses": {
          "200": {
            "description": "The user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user",
        "description": "Also removes the user's favorites. Their inquiries are kept without a user link.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1a2b3c0000abcd1234efgh"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "priya.sharma@example.com"
          },
          "fullName": {
            "type": "string",
            "example": "Priya Sharma"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+91 98765 43210"
          },
          "role": {
            "type": "string",
            "enum": [
              "BUYER",
              "AGENT",
              "ADMIN"
            ],
            "example": "BUYER"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PaginatedUsers": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
      "CreateUser": {
        "type": "object",
        "required": [
          "email",
          "fullName"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "priya.sharma@example.com"
          },
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "example": "Priya Sharma"
          },
          "phone": {
            "type": "string",
            "example": "+91 98765 43210"
          },
          "role": {
            "type": "string",
            "enum": [
              "BUYER",
              "AGENT",
              "ADMIN"
            ],
            "default": "BUYER"
          }
        }
      },
      "UpdateUser": {
        "type": "object",
        "description": "Send only the fields you want to change",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "BUYER",
              "AGENT",
              "ADMIN"
            ]
          }
        },
        "example": {
          "fullName": "Priya S. Sharma",
          "phone": "+91 90000 11111"
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 404
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Resource not found"
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 57
          },
          "totalPages": {
            "type": "integer",
            "example": 3
          }
        }
      },
...
    },
    "parameters": {
      "Page": {
        "in": "query",
        "name": "page",
        "required": false,
        "description": "Page number, starting at 1",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "Limit": {
        "in": "query",
        "name": "limit",
        "required": false,
        "description": "Items per page. Defaults to PAGINATION_DEFAULT_LIMIT, capped at PAGINATION_MAX_LIMIT.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "example": 20
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input. `message` lists every problem.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 400,
              "message": [
                "email must be an email"
              ],
              "error": "Bad Request"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "A record with the same unique value already exists",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded (RATE_LIMIT_MAX requests per RATE_LIMIT_TTL_SECONDS)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "The database is not reachable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 503,
              "message": "The database is not reachable",
              "error": "Service Unavailable"
            }
          }
        }
      }
    }
  },
...
}